backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.8.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
channels==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.27.2
gunicorn==20.0.4
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msgpack==0.6.2
phonenumbers==8.11.5
promise==2.3
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-decouple==3.3
python-http-client==3.2.6
pytz==2019.3
requests==2.23.0
rsa==4.0
Rx==1.6.1
sendgrid==6.1.3
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
twilio==6.37.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
zope.interface==4.6.0
